GVT-3016: Refactor layout alignment map feature  z-indexes, styles and feature creation
