+++ ui/src/map/layers/alignment/reference-line-alignment-layer.ts
@@ -8,11 +8,13 @@ import { createLayer, GeoviiteMapLayer, loadLayerData } from 'map/layers/utils/l
 import { deduplicate, filterNotEmpty } from 'utils/array-utils';
 import {
     getReferenceLineMapAlignmentsByTiles,
+    LayoutAlignmentDataHolder,
     ReferenceLineAlignmentDataHolder,
 } from 'track-layout/layout-map-api';
 import {
-    createAlignmentFeatures,
+    createAlignmentFeature,
     findMatchingAlignments,
+    isHighlighted,
     NORMAL_ALIGNMENT_OPACITY,
     OTHER_ALIGNMENTS_OPACITY_WHILE_SPLITTING,
 } from 'map/layers/utils/alignment-layer-utils';
@@ -20,6 +22,7 @@ import { ReferenceLineId } from 'track-layout/track-layout-model';
 import { Rectangle } from 'model/geometry';
 import { Stroke, Style } from 'ol/style';
 import mapStyles from 'map/map.module.scss';
+import Feature from 'ol/Feature';
 
 let shownReferenceLinesCompare: string;
 
@@ -38,6 +41,20 @@ const referenceLineStyle = new Style({
     }),
     zIndex: 0,
 });
+
+export function createAlignmentFeatures(
+    alignments: LayoutAlignmentDataHolder[],
+    selection: Selection,
+    showEndTicks: boolean,
+): Feature<LineString | OlPoint>[] {
+    return alignments.flatMap((alignment) => {
+        const highlighted = isHighlighted(selection, alignment.header);
+        const style = highlighted ? [highlightedReferenceLineStyle] : [referenceLineStyle];
+        const endTickStyle = highlighted ? highlightedReferenceLineStyle : referenceLineStyle;
+
+        return createAlignmentFeature(alignment, style, showEndTicks ? endTickStyle : undefined);
+    });
+}
 
 const layerName: MapLayerName = 'reference-line-alignment-layer';
 
@@ -70,13 +87,7 @@ export function createReferenceLineAlignmentLayer(
         getReferenceLineMapAlignmentsByTiles(changeTimes, mapTiles, layoutContext);
 
     const createFeatures = (referenceLines: ReferenceLineAlignmentDataHolder[]) =>
-        createAlignmentFeatures(
-            referenceLines,
-            selection,
-            false,
-            referenceLineStyle,
-            highlightedReferenceLineStyle,
-        );
+        createAlignmentFeatures(referenceLines, selection, false);
 
     const onLoadingChange = (
         loading: boolean,
